Generalize and improve analyzer/chaotic iteration.
Chaotic iteration now finds the starting node in the cfg instead of
assuming label 0 (this used to be in a TODO). Its return value has also
been modified from a list of abstract lattice elements into a dictionary
from the labels to the lattice elements, in consideration of cases where
the labels aren't necessarily consequtive.