Revert analysis interface to its initial version.
This means that BaseAnalysis now contains all lattice methods such as
top, bottom, equiv, etc.

A new abstract class LatticeBasedAnalysis is created in lattice.py to
allow for creating analyses based on a lattice from the Lattice interface
with ease by delegeting the aforementioned lattice-methods in the
analysis to a nested Lattice object.